Example: Inject an object into other object using xml-file
The objective here is to see how one configures the Spring xml-file
property tag for injecting an object into another object.

XML-file Syntax:
<bean ....>
  <property name="..." ref="xxx"/>
</bean>

Compare to the syntax for injecting a string:
<bean ....>
  <property name="..." value="xxx was here"/>
</bean>

Add a new class implemeting the PurchasingService interface as well as
update the test client and verify the implementation works. Note, not
any unit test created, just a simple test to see that buyBook() works.