Tell Spring to handle a 'bean' as a 'stateful bean'
Normally /Spring/ only creates one instance of a /service object/. The
reasons for this is that often the object data is directly stored in a
database entry and not in an object in memory.

For the cases where you want to store /client/ data as separate
objects in memory (e.g. 'shopping chart') it can be accomplished
by configuring /Spring/ to handle a bean as a /Stateful Bean/.