Inject using constructor arguments instead of using setters
An alternative way to inject references into an object is to use
object constructor arguments instead of using set-methods.

This requires a different syntax in the Spring container xml-file.

We use the /constructor-arg/ tag instead of the /property/ tag.

Note that if the constructor takes multiple arguments of the same type
additional information must be provided in the /constructor-arg/ tag
for Spring to be able to know what goes where.